TemporarySynchronizationContext をローカルクラスに格下げ
