--- Assets/UniGLTF/Runtime/UniGLTF/IO/TemporarySynchronizationContext.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 62f8f380aadfc284f9bdc2d58f84bf48
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
